refactor(static-analysis): extract MCP capability detection into CapabilityDetector
native_analyzer.py grew to ~4.3k lines spanning three distinct concerns:
generic AST extraction, dataflow/taint analysis, and MCP-capability
detection. The capability-detection slice is roughly 1.7k lines and
operates at a different level of abstraction than the rest of the
module, which made the file hard to review, test, and reason about.

This change pulls the capability-detection slice into a sibling module,
``mcpscanner.core.static_analysis.capability_detector``, while keeping
``NativeAnalyzer`` source-compatible:

* New ``CapabilityDetector`` wraps a ``NativeAnalyzer`` and exposes both
  the legacy ``extract_mcp_capability_contexts()`` (returning
  ``FunctionContext``) and a new ``detect()`` method that returns
  lightweight ``CapabilityRecord`` dataclasses suited for downstream
  consumers that don't need full taint context.
* All MCP-only constants, regexes, decorator/registration helpers, and
  Python/Tree-sitter walkers move with it. ``native_analyzer.py`` keeps
  the byte-level prefilter (``_has_mcp_markers`` + cache) so directory
  scanners can still skip non-MCP files cheaply, and re-exports the
  migrated symbols for backward compatibility.
* ``NativeAnalyzer.extract_mcp_capability_contexts`` is now a thin shim
  that delegates to ``CapabilityDetector`` and emits a
  ``DeprecationWarning`` so existing integrations keep working while
  they migrate.
* ``BehavioralCodeAnalyzer._analyze_source_code`` calls
  ``CapabilityDetector`` directly to avoid spamming the deprecation
  warning on every behavioral scan; the ``FunctionContext`` shape it
  consumes is unchanged.
* ``mcpscanner.core.static_analysis`` re-exports ``CapabilityDetector``
  and ``CapabilityRecord`` as part of the public API.

Why this layout:
- Each module now has a single job; ``native_analyzer.py`` drops from
  ~4.3k to ~1.6k lines (well under the 2.2k acceptance bar) and is
  focused on AST/dataflow concerns.
- Capability detection becomes independently testable and easier to
  evolve (e.g. adding new SDK registration shapes) without touching
  taint-analysis code.
- Caches that need to outlive a single detector pass (prefilter,
  function index, annotation index) are stored on the ``NativeAnalyzer``
  instance, so behavior under repeated scans of the same file is
  unchanged.

Tested:
- ``pytest tests/static_analysis -q``: 118 passed, 6 skipped.
- ``pytest tests/test_behavioral*.py -q``: 79 passed.
- Full suite minus a known-flaky network-bound OpenAI integration test:
  835 passed, 7 skipped.
- New ``tests/static_analysis/test_capability_detector.py`` adds 17
  unit tests covering ``CapabilityDetector.detect()`` across Python /
  TS / Java fixtures, ``CapabilityRecord`` shape, ``__getattr__``
  delegation to the wrapped analyzer, the deprecation shim, and the
  re-exports / package-level public API.
